chore: renamed utils/validate to helpers/validate
