+++ src/core/client/framework/lib/validation.tsx
@@ -7,7 +7,7 @@ import {
   USERNAME_MAX_LENGTH,
   USERNAME_MIN_LENGTH,
   USERNAME_REGEX,
-} from "talk-common/utils/validate";
+} from "talk-common/helpers/validate";
 
 import {
   EMAILS_DO_NOT_MATCH,
