+++ src/core/common/helpers/validate.ts
